+++ src/versioned-entity/manager.ts
@@ -65,7 +65,8 @@ type VersionedEntityUpdateOptions<
 export class VersionedEntityManager<
   T extends FindReplaceTransaction,
   E extends Event<string, VersionedEntity>,
-> extends VersionedEntityEventProcessor<T, E, EventData<E>> {
+  R extends TransactionRunner<T> = TransactionRunner<T>,
+> extends VersionedEntityEventProcessor<T, E, EventData<E>, R> {
   /**
    * Creates a new {@link VersionedEntityManager}.
    *
@@ -78,9 +79,9 @@ export class VersionedEntityManager<
     readonly topic: string,
     readonly eventType: { new (): E },
     entityType: { new (): EventData<E> },
-    runner: TransactionRunner<T>,
+    runner: R,
   ) {
-    super(entityType, ({ data }) => data, runner);
+    super(entityType, async ({ data }) => data, runner);
   }
 
   /**
@@ -155,7 +156,7 @@ export class VersionedEntityManager<
    * @param options Options when finding the entity.
    * @returns The found entity.
    */
-  protected async findExistingEntityOrFail(
+  protected async findExistingEntityWithVersionOrFail(
     entity: Partial<EventData<E>>,
     knownUpdatedAt: Date,
     options: {
@@ -327,7 +328,7 @@ export class VersionedEntityManager<
     return await this.runner.runInNewOrExisting(
       options.transaction,
       async (transaction) => {
-        const existingEntity = await this.findExistingEntityOrFail(
+        const existingEntity = await this.findExistingEntityWithVersionOrFail(
           update as Partial<EventData<E>>,
           knownUpdatedAt,
           { transaction, existingEntity: options.existingEntity },
@@ -374,7 +375,7 @@ export class VersionedEntityManager<
     return await this.runner.runInNewOrExisting(
       options.transaction,
       async (transaction) => {
-        const existingEntity = await this.findExistingEntityOrFail(
+        const existingEntity = await this.findExistingEntityWithVersionOrFail(
           entityKey,
           knownUpdatedAt,
           { transaction, existingEntity: options.existingEntity },